# docs/swagger.yaml
openapi: 3.0.3
info:
  title: API Gateway Service
  description: |
    Multi-channel messaging API Gateway that provides unified access to SMS, Email, and WhatsApp messaging services.
    
    ## Features
    - Multi-channel message sending (SMS, Email, WhatsApp)
    - Bulk message processing
    - Real-time webhook handling
    - Rate limiting and authentication
    - Conversation management
    - Channel health monitoring
    
    ## Authentication
    Supports two authentication methods:
    - **JWT Bearer Token**: For user-based authentication
    - **API Key**: For service-to-service authentication (X-API-Key header)
    
    ## Rate Limiting
    Hierarchical rate limiting is applied:
    - Global: 1000 requests/second
    - Per-tenant: 50 requests/second
    - Per-channel: Channel-specific limits
    
  version: 1.0.0
  contact:
    name: API Gateway Team
    email: api-gateway@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-gateway.company.com/api/v1
    description: Production server
  - url: https://staging-api-gateway.company.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /messages/send:
    post:
      tags:
        - Messages
      summary: Send a message
      description: |
        Send a single message through one or multiple channels.
        Supports SMS, Email, and WhatsApp channels with channel-specific content formatting.
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              sms_message:
                summary: SMS Message
                value:
                  to: "+1234567890"
                  channel: "sms"
                  content:
                    text: "Hello! Your order #12345 has been shipped."
                  metadata:
                    orderId: "12345"
              email_message:
                summary: Email Message
                value:
                  to: "user@example.com"
                  channel: "email"
                  content:
                    subject: "Order Confirmation"
                    text: "Thank you for your order!"
                    html: "<h1>Thank you for your order!</h1><p>Order #12345</p>"
                  conversationId: "conv_123"
              whatsapp_message:
                summary: WhatsApp Message
                value:
                  to: "+1234567890"
                  channel: "whatsapp"
                  content:
                    text: "Your appointment is confirmed for tomorrow at 2 PM"
                    templateId: "appointment_confirmation"
                    templateParams:
                      appointment_time: "2 PM"
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /messages/bulk:
    post:
      tags:
        - Messages
      summary: Send bulk messages
      description: |
        Send messages to multiple recipients efficiently.
        Supports up to 10,000 recipients per batch with automatic chunking.
      operationId: sendBulkMessages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMessageRequest'
      responses:
        '200':
          description: Bulk messages queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkMessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Get message status
      description: Retrieve the current status and details of a specific message
      operationId: getMessageStatus
      parameters:
        - name: messageId
          in: path
          required: true
          description: Unique message identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{messageId}/events:
    get:
      tags:
        - Messages
      summary: Get message events
      description: Retrieve all events (sent, delivered, failed, etc.) for a specific message
      operationId: getMessageEvents
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of events to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Message events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEvents'

  # Channel APIs
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
  schemas:
...
    SendMessageRequest:
      type: object
      required:
        - to
        - channel
        - content
      properties:
        to:
          oneOf:
            - type: string
              description: Single recipient
            - type: array
              items:
                type: string
              description: Multiple recipients
        channel:
          type: string
          enum: [sms, email, whatsapp]
          description: Messaging channel
        content:
          $ref: '#/components/schemas/MessageContent'
        conversationId:
          type: string
          format: uuid
          description: Optional conversation ID to group related messages
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for the message
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          description: Message priority level
        scheduledAt:
          type: string
          format: date-time
          description: Schedule message for future delivery
    MessageContent:
      type: object
      properties:
        text:
          type: string
          maxLength: 1600
          description: Plain text content (required for SMS, optional for email)
        subject:
          type: string
          maxLength: 200
          description: Email subject line (required for email)
        html:
          type: string
          maxLength: 100000
          description: HTML content for email
        templateId:
          type: string
          description: Template ID for WhatsApp
        templateParams:
          type: object
          additionalProperties: true
          description: Parameters for template substitution
    BulkMessageRequest:
      type: object
      required:
        - recipients
        - channels
        - content
      properties:
        recipients:
          oneOf:
            - type: array
              items:
                type: string
              maxItems: 10000
              description: List of recipient addresses
            - type: object
              properties:
                type:
                  type: string
                  enum: [file, segment]
                source:
                  type: string
                  description: File path or segment ID
                format:
                  type: string
                  enum: [csv, json]
        channels:
          type: array
          items:
            type: string
            enum: [sms, email, whatsapp]
          minItems: 1
          maxItems: 3
        content:
          $ref: '#/components/schemas/MessageContent'
        options:
          type: object
          properties:
            batchSize:
              type: integer
              minimum: 100
              maximum: 1000
              default: 500
            rateLimit:
              type: integer
              minimum: 1
              maximum: 50
              default: 10
            scheduledAt:
              type: string
              format: date-time
            retryFailures:
              type: boolean
              default: true
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            messageId:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, sent, delivered, failed]
            channel:
              type: string
              enum: [sms, email, whatsapp]
            queuedAt:
              type: string
              format: date-time
            estimatedDelivery:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    BulkMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            batchId:
              type: string
              format: uuid
            totalRecipients:
              type: integer
            estimatedDuration:
              type: string
            status:
              type: string
              enum: [queued, processing, completed, failed]
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    MessageStatus:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            messageId:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, sent, delivered, failed, expired]
            channel:
              type: string
            recipient:
              type: string
            createdAt:
              type: string
              format: date-time
            sentAt:
              type: string
              format: date-time
            deliveredAt:
              type: string
              format: date-time
            failedAt:
              type: string
              format: date-time
            errorCode:
              type: string
            errorMessage:
              type: string
            retryCount:
              type: integer
            metadata:
              type: object
    MessageEvents:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            messageId:
              type: string
              format: uuid
            events:
              type: array
              items:
                type: object
                properties:
                  eventId:
                    type: string
                    format: uuid
                  type:
                    type: string
                    enum: [queued, sent, delivered, failed, clicked, opened]
                  timestamp:
                    type: string
                    format: date-time
                  details:
                    type: object
            totalEvents:
              type: integer
            hasMore:
              type: boolean

    # Channel Schemas
...
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
          description: Unique request identifier for tracing
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          description: API version
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
        meta:
          $ref: '#/components/schemas/ResponseMeta'
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: VALIDATION_ERROR
                      message:
                        example: Request validation failed
                      details:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                            message:
                              type: string
                            value:
                              type: string
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: UNAUTHORIZED
                      message:
                        example: Invalid or missing authentication token
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: string
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: string
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: string
          description: Time when the rate limit resets (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        example: Too many requests, please retry after 30 seconds
                      details:
                        properties:
                          retryAfter:
                            type: integer
                          resetAt:
                            type: string
                            format: date-time
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: NOT_FOUND
                      message:
                        example: The requested resource was not found
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: SERVICE_UNAVAILABLE
                      message:
                        example: Service temporarily unavailable
...